Improves https://github.com/brillout/vite-plugin-ssr/issues/1115 by removing try/catch to avoid error handling overhead when pathname is relative and no `base` is passed to URL constructor.

Benchmark using the `npm create vite-plugin-ssr starter` repo on the index page. Used [bombardier](https://github.com/codesenberg/bombardier) on Macbook with M1 Pro.

before:
```
Statistics        Avg      Stdev        Max
  Reqs/sec      1312.95     223.14    3744.83
  Latency       94.75ms    17.55ms   244.31ms
```
after:
```
Statistics        Avg      Stdev        Max
  Reqs/sec      2269.03     439.82    6707.62
  Latency       54.98ms    12.65ms   190.80ms
```