@brillout 2023-09-16
> before:
> 
> ```
> Statistics        Avg      Stdev        Max
>   Reqs/sec      1312.95     223.14    3744.83
>   Latency       94.75ms    17.55ms   244.31ms
> ```
> 
> after:
> 
> ```
> Statistics        Avg      Stdev        Max
>   Reqs/sec      2269.03     439.82    6707.62
>   Latency       54.98ms    12.65ms   190.80ms
> ```

That's very surprising. I wonder why `new URL()` is so slow.