@dajinchu 2023-09-16
> That's very surprising. I wonder why new URL() is so slow.

Its less that URL is slow and more that error handling is slow