@brillout 2023-09-16
> Its less that URL is slow and more that error handling is slow

I would find this even more surprising.

I guess the way to find out is to benchmark the difference between no `new URL()` usage at all and using it once per `parseUrl()` call 👀.